# rentalcar-api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Car Server",
    "description": "Api documentation rental car server",
    "version": "1.0.0",
    "contact": {
      "name": "Rafli Andreansyah",
      "email": "rafliandreansyah957@gmail.com",
      "url": "https://www.linkedin.com/in/rafli-andreansyah-86299217a"
    }
  },
  "servers": [
    {
      "url": "https://{env}:{port}/api/{version}",
      "description": "Server developer rental car",
      "variables": {
        "env": {
          "description": "Environtment url",
          "default": "localhost",
          "enum": [
            "localhost"
          ]
        },
        "port": {
          "description": "port server rental car",
          "default": "8080",
          "enum": [
            "8080",
            "8081"
          ]
        },
        "version": {
          "description": "api version",
          "default": "v1",
          "enum": [
            "v1"
          ]
        }
      }
    }
  ],
  "paths": {
...
    "/cars": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "get list of cars",
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transmission",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date_rent",
            "in": "query",
            "description": "required for role user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "description": "required for role user",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order_by_date_created",
            "in": "query",
            "description": "only admin can access",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactions",
            "in": "query",
            "description": "onlu admin can access",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "1"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "20"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success get list car",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "total_item": 50,
                      "per_page": 10,
                      "current_page": 1,
                      "last_page": 5,
                      "next_page_url": null,
                      "prev_page_url": null,
                      "status": 200,
                      "error": null,
                      "data": [
                        {
                          "id": "00FYY-66DYY-666DH",
                          "name": "Toyota Fortuner",
                          "image_url": "https://fortunerfjjs888s.com",
                          "year": 2020,
                          "transmission": "Automatic",
                          "price": 250000.00
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "create data cars and only admin can create",
        "tags": [
          "Cars"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RequestPostAndPutCars"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success create car",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 201,
                      "error": null,
                      "data": {
                        "id": "0KDD2-HJSK9-KKS00-23KSS",
                        "name": "Pajero",
                        "image": "https://pajero.com",
                        "image_detail": [
                          {
                            "id": "DD23S-KDSKJ-99DUS",
                            "image": "https://pajero1.com"
                          }
                        ],
                        "brand": "mitsubishi",
                        "year": 2022,
                        "capacity": 7,
                        "cc": 2000,
                        "price": 300000.00,
                        "description": "this is description car",
                        "transmission": "automatic",
                        "tax": 0,
                        "discount": 0,
                        "is_active": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/cars/{id}": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "get car data by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Cars"
        ],
        "responses": {
          "200": {
            "description": "success get data car",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": {
                        "id": "00JS8-77SHS-88JJJ-UHH77-2DSSA",
                        "name": "Avanza",
                        "image": "https://avanza.com",
                        "brand": "Toyota",
                        "year": 2020,
                        "capacity": 6,
                        "cc": 1500,
                        "price_per_day": 250000.00,
                        "tax": null,
                        "discount": null,
                        "description": "description of avanza",
                        "transmission": "automatic",
                        "is_active": true,
                        "has_authorization": [
                          {
                            "id": "0092HSH-0092HSH",
                            "image_url": "https://user-profilerafli.com",
                            "email": "rafli@gmail.com",
                            "name": "rafli a",
                            "phone": "+62039477755",
                            "is_active": "true"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/CarNotFoundError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "delete car data by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Cars"
        ],
        "responses": {
          "200": {
            "description": "success delete car data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataString"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": "success delete car"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/CarNotFoundError"
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "patch": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "edit car data by id",
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RequestPostAndPutCars"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success edit cars",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": {
                        "id": "0KDD2-HJSK9-KKS00-23KSS",
                        "name": "Pajero",
                        "image": "https://pajero.com",
                        "image_detail": [
                          {
                            "id": "DD23S-KDSKJ-99DUS",
                            "image": "https://pajero1.com"
                          }
                        ],
                        "brand": "mitsubishi",
                        "year": 2022,
                        "capacity": 7,
                        "cc": 2000,
                        "price": 300000.00,
                        "description": "this is description car",
                        "transmission": "automatic",
                        "tax": 0,
                        "discount": 0,
                        "is_active": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/CarNotFoundError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/cars/authorization": {
      "get": {
        "security": [
          {
            "AppSecurity": []
          }
        ],
        "description": "get authorization car",
        "tags": [
          "Cars"
        ],
        "responses": {
          "200": {
            "description": "success get list authorization car",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
                },
                "examples": {
                  "success": {
                    "value": {
                      "total_item": 50,
                      "per_page": 10,
                      "current_page": 1,
                      "last_page": 5,
                      "next_page_url": null,
                      "prev_page_url": null,
                      "status": 200,
                      "error": null,
                      "data": [
                        {
                          "id": "00FYY-66DYY-666DH",
                          "name": "Toyota Fortuner",
                          "image_url": "https://fortunerfjjs888s.com",
                          "year": 2020,
                          "transmission": "Automatic",
                          "price": 250000.00,
                          "has_authorizations": [
                            {
                              "id": "0092HSH-0092HSH",
                              "image_url": "https://user-profilerafli.com",
                              "name": "rafli a",
                              "email": "rafli@gmail.com",
                              "phone": "+62039477755",
                              "is_active": "true"
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessForbiddenErrorPaging"
          }
        }
      },
      "post": {
        "description": "post authorization car",
        "security": [
          {
            "AppSercurity": []
          }
        ],
        "tags": [
          "Cars"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "user_id": {
                    "type": "array",
                    "required": true,
                    "items": {
                      "type": "string"
                    }
                  },
                  "car_id": {
                    "type": "array",
                    "required": true,
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "examples": {
                "success": {
                  "value": {
                    "user_id": [
                      "KD009-DWK9D-DMWWW-9DWMKDMW",
                      "aD009-DWK9D-DMWWW-9DWMKDMW",
                      "mD009-DWK9D-DMWWW-9DWMKDMW"
                    ],
                    "car_id": [
                      "KD049-DWK9D-DMWWW-9DWMKDMW",
                      "KD209-DWK9D-DMWWW-9DWMKDMW",
                      "KD109-DWK9D-DMWWW-9DWMKDMW"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success post authorization car",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralResponseWithDataString"
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "error": null,
                      "data": "success add car authorization"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AppSecurity": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GeneralResponseWithDataObject": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "GeneralResponseWithDataArrayObject": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "GeneralResponseWithDataString": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "RequestPostAndPutCars": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true
          },
          "image": {
            "type": "array",
            "required": true,
            "items": {
              "format": "binary",
              "type": "string"
            }
          },
          "image_detail": {
            "type": "array",
            "required": false,
            "items": {
              "format": "binary",
              "type": "string"
            }
          },
          "brand": {
            "type": "string",
            "required": true
          },
          "year": {
            "type": "number",
            "required": true
          },
          "capacity": {
            "type": "number",
            "required": true
          },
          "cc": {
            "type": "number",
            "required": true
          },
          "price": {
            "type": "number",
            "required": true
          },
          "description": {
            "type": "string",
            "required": true
          },
          "transmission": {
            "type": "string",
            "required": true
          },
          "tax": {
            "type": "number",
            "required": false
          },
          "discount": {
            "type": "number",
            "required": false
          }
        }
      }
    },
    "responses": {
...
      "CarNotFoundError": {
        "description": "Response error for car not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataObject"
            },
            "examples": {
              "error": {
                "value": {
                  "status": 404,
                  "error": "car doesn't exist",
                  "data": null
                }
              }
            }
          }
        }
      },
...
      "AccessForbiddenError": {
        "description": "response error forbidden access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
            },
            "examples": {
              "forbidden": {
                "value": {
                  "status": 403,
                  "error": "access forbidden to access this data",
                  "data": null
                }
              }
            }
          }
        }
      },
      "AccessForbiddenErrorPaging": {
        "description": "response error forbidden access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
            },
            "examples": {
              "forbidden": {
                "value": {
                  "total_item": 50,
                  "per_page": 10,
                  "current_page": 1,
                  "last_page": 5,
                  "next_page_url": null,
                  "prev_page_url": null,
                  "status": 403,
                  "error": "access forbidden to access this data",
                  "data": null
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "response error forbidden access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralResponseWithDataArrayObject"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "status": 401,
                  "error": "authorization required",
                  "data": null
                }
              }
            }
          }
        }
      }
    }
  }
}